Migrate away from legacy SLVS binding
The SLVS (legacy) binding used to make changes in the project files that later were commited to the repo. This is no
longer the case, the only file related to the binding that is commited to repo is the shared configuration. SLVS provides
a migration wizard that undo the legacy binding changes. This commit is result of running the migration wizard.